Temperature working properly

+++ measurement/temperature.go
@@ -1,14 +1,14 @@
 package measurement
 
-const TemperatureDeltaStrength = 0.2
+const TemperatureDeltaStrength = float32(0.2)
 
 type Temperature struct {
 	DegreesCelsius float32
 }
 
 func (t *Temperature) ApplyTempAffect(temperature Temperature, seconds float32) {
-	temperatureDeltaNum := temperature.DegreesCelsius - t.DegreesCelsius
-	t.DegreesCelsius = t.DegreesCelsius + temperatureDeltaNum*(TemperatureDeltaStrength*seconds)
+	temperatureDeltaNum := float32(temperature.DegreesCelsius - t.DegreesCelsius)
+	t.DegreesCelsius = t.DegreesCelsius + (temperatureDeltaNum * (TemperatureDeltaStrength * seconds))
 }
 
 func (t *Temperature) ApplyTempDeltaByDegreesC(degreesCelsius float32, seconds float32) {
